General hash table code and HF1 implemented
This hash function just returns stringData's bytes interpreted numerically instead of as a string. I chose to start with this approach because this is about as simple as hashing algorithms get and so it lets me establish a good baseline of performance. The title table had 0 unused slots, 7009 collisions, and a construction time of .06746s. The quote table had 0 unused slots, 5603 collisions, and a construction time of .06795s. Interestingly, raising the table size above the minimum value of 15,001 needed to hold all the data resulted in more collisions, not less.