Implemented HF2 and fixed minor errors in hash table logic
For this implementation, I tried bisecting and swapping the bytes of stringData before interpreting it as a number. I chose this approach so that the lower bits that most-directly affect the index generated by modding would be more random and not conform to common English word-ending patterns. The title table had 0 unused slots, 7086 collisions, and a construction time of .11642s. The quote table had 0 unused slots, 5554 collisions, and a construction time of .15331s. Like before, increasing tableSize unexpectedly increased the collision counts