Implemented HF3
For this implementation, instead of just somehow reading stringData directly as a number, I first squared it to amplify the given stringData's numerical difference from other, similar stringDatas, and from that new number I pulled out a key from the central bits, which tend to be the most random. I chose this approach because simply interpreting stringData as an int, even in a somewhat random order as in HF2, didn't really improve collision statistics, so my hope was that by squaring int(stringData) I could get slightly different stringData values to produce very different results and thereby decrease collisions. The title table had 0 unused slots, 7058 collisions, and a construction time of .13817s. The quote table had 0 unused slots, 5549 collisions, and a construction time of 1.19516s. This relatively long construction time on the latter is likely a result of how long the quotes are compared to the titles, a length difference further exaggerated by the squaring operation. Both collision numbers here are slightly better than the HF2 ones, but compared to HF1 only the quote table shows improvement.