HF5 Implementation
In this implementation, I read stringData as a big integer and modded it by roughly tableSize*6 to get the key. I chose this implementation over trying to add extra layers of randomness through nested mathematical operations because the latter approach yielded little fruit in the past, so I instead wanted to focus on just making sure one end of the allowed index range wasn't being favored over the other. The title table had 0 unused slots, 6983 collisions, and a construction time of .06789s. The quote table had 0 unused slots, 5499 collisions, and a construction time of .06231s. The collision numbers for both of the tables here are better than in any other hash function implementation yet performed.